requirements.txt
appdirs==1.4.3
asgiref==3.2.7
certifi==2019.11.28
chardet==3.0.4
distlib==0.3.0
dj-database-url==0.5.0
Django==3.0.4
django-heroku==0.3.1
djangorestframework==3.11.0
drf-nested-routers==0.91
filelock==3.0.12
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.6.0
importlib-resources==1.5.0
psycopg2==2.8.4
psycopg2-binary==2.8.5
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
virtualenv==20.0.18
whitenoise==5.0.1
zipp==3.1.0
